Corrected `ushort` to `uint` on the tags of specialized `Put` methods.

+++ Src/IronJS.Runtime/Objects/CommonObject.cs
@@ -724,7 +724,7 @@ namespace IronJS.Runtime
         // Put methods for setting indexes to doubles
         //----------------------------------------------------------------------------
 
-        public void Put(BoxedValue index, object value, ushort tag)
+        public void Put(BoxedValue index, object value, uint tag)
         {
             uint i = 0;
             if (TypeConverter.TryToIndex(index, out i))
@@ -737,12 +737,12 @@ namespace IronJS.Runtime
             }
         }
 
-        public void Put(bool index, object value, ushort tag)
+        public void Put(bool index, object value, uint tag)
         {
             this.Put(TypeConverter.ToString(index), value, tag);
         }
 
-        public void Put(double index, object value, ushort tag)
+        public void Put(double index, object value, uint tag)
         {
             uint parsed = 0;
             if (TypeConverter.TryToIndex(index, out parsed))
@@ -755,7 +755,7 @@ namespace IronJS.Runtime
             }
         }
 
-        public void Put(object index, object value, ushort tag)
+        public void Put(object index, object value, uint tag)
         {
             var s = TypeConverter.ToString(index);
             var parsed = 0u;
@@ -770,12 +770,12 @@ namespace IronJS.Runtime
             }
         }
 
-        public void Put(Undefined index, object value, ushort tag)
+        public void Put(Undefined index, object value, uint tag)
         {
             this.Put("undefined", value, tag);
         }
 
-        public void Put(CommonObject index, object value, ushort tag)
+        public void Put(CommonObject index, object value, uint tag)
         {
             string s = TypeConverter.ToString(index);
             uint parsed = 0;
